Reformated README to match standard-readme
I have tried to make the README a bit easier to use. While not perfect, I think that the process is a bit clenaer here. Instead of the standard 'Usage' section, I've opted for a 'Development' section, and I think that as a whole the README is better for it. I've also updated the Contribute section, added a TOC, and made the installation a bit clearer - before, it was hard to disambiguate developing on Orbit from using it.

Let me know what you think! This was a bit of work, but I think it is a better README now, and I would be curious to discuss all of these changes.